Move hearing controls lower
