+++ VideoWeb/VideoWeb/ClientApp/src/app/services/events.service.ts
@@ -4,7 +4,8 @@ import { AdalService } from 'adal-angular4';
 import { Observable, Subject } from 'rxjs';
 import { ErrorService } from 'src/app/services/error.service';
 import { Heartbeat } from '../shared/models/heartbeat';
-import { ParticipantMediaStatus, ParticipantMediaStatusMessage } from '../shared/models/participant-media-status';
+import { ParticipantMediaStatus } from '../shared/models/participant-media-status';
+import { ParticipantMediaStatusMessage } from '../shared/models/participant-media-status-message';
 import { ConfigService } from './api/config.service';
 import { ConferenceStatus, ConsultationAnswer, EndpointStatus, ParticipantStatus, RoomType } from './clients/api-client';
 import { Logger } from './logging/logger-base';
