+++ VideoWeb/VideoWeb/ClientApp/src/app/shared/models/participant-media-status-message.ts
@@ -1,0 +1,5 @@
+import { ParticipantMediaStatus } from './participant-media-status';
+
+export class ParticipantMediaStatusMessage {
+    constructor(public conferenceId: string, public participantId: string, public mediaStatus: ParticipantMediaStatus) {}
+}
