+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/hearing-controls/hearing-controls.component.html
@@ -12,7 +12,7 @@
         Your browser does not support Video
       </video>
     </div>
-    <div class="controls-container govuk-grid-column-one-third govuk-!-margin-top-9" [ngClass]="{ 'judge-controls-padding': isJudge }">
+    <div class="controls-container govuk-grid-column-one-third" [ngClass]="{ 'judge-controls-padding': isJudge }">
       <div class="icons-container">
         <img
           id="toggle-self-view-img"
